database connectivity part 1

+++ order.h
@@ -11,7 +11,7 @@ class Order {
 
     char order_type;
     std::string orderNumber;
-    std::vector<std::tuple<int, std::string, int>> order_items;
+    std::vector<std::tuple<int, std::string, double>> order_items;
     Customer this_customer;
     Menu this_menu;
 
@@ -20,7 +20,6 @@ public:
     // Constructor
     Order(char new_order_type, std::string new_orderNumber, Customer this_customer, Menu this_menu);
 
-    bool add_items(Order&, std::vector<std::tuple<int, std::string, int>>);
-    //void remove_item(Order&, int&, std::vector<std::tuple<int, std::string, int>>);
+    bool add_items(Order&, std::vector<std::tuple<int, std::string, double>>);
     void print_order(Customer);
 };
